+++ trunk/src/OpenMesh/Core/Mesh/BaseKernel.hh
@@ -630,10 +630,10 @@ protected: // low-level access non-public
 public: //----------------------------------------------------- element numbers
 
 
-  virtual uint n_vertices()  const { return 0; }
-  virtual uint n_halfedges() const { return 0; }
-  virtual uint n_edges()     const { return 0; }
-  virtual uint n_faces()     const { return 0; }
+  virtual size_t n_vertices()  const { return 0; }
+  virtual size_t n_halfedges() const { return 0; }
+  virtual size_t n_edges()     const { return 0; }
+  virtual size_t n_faces()     const { return 0; }
 
 
 protected: //------------------------------------------- synchronize properties
